Initial Commit
We've got to get started somewhere!  Format, files, libraries, and everything is subject to change.  At this point, this is mainly to work as a starting point and to help lead to discussions of what we need to gather, how we need to organize it, and how it'd be the easiest to implement and make use of in the future.